+++ example/shared/src/components/sheets/PromptSheet.tsx
@@ -46,7 +46,6 @@ export const PromptSheet = forwardRef((props: PromptSheetProps, ref: Ref<TrueShe
       ref={sheetRef}
       name="prompt-sheet"
       detents={[0.75, 1]}
-      scrollable
       scrollableOptions={{
         keyboardScrollOffset: FOOTER_HEIGHT + SPACING,
         topScrollEdgeEffect: 'soft',