+++ ios/TrueSheetContainerView.mm
@@ -26,7 +26,6 @@
 
 #import <React/RCTConversions.h>
 #import <React/RCTLog.h>
-#import <react/renderer/core/LayoutMetrics.h>
 
 using namespace facebook::react;
 
@@ -56,7 +55,6 @@ using namespace facebook::react;
   TrueSheetFooterView *_footerView;
   TrueSheetPeekView *__weak _peekView;
   TrueSheetKeyboardObserver *_keyboardObserver;
-  BOOL _scrollableSet;
 }
 
 #pragma mark - Initialization
@@ -74,7 +72,6 @@ using namespace facebook::react;
     _contentView = nil;
     _headerView = nil;
     _footerView = nil;
-    _scrollableSet = NO;
     self.isAccessibilityElement = NO;
   }
   return self;
@@ -109,11 +106,6 @@ using namespace facebook::react;
 
 #pragma mark - Layout
 
-- (void)layoutSubviews {
-  [super layoutSubviews];
-  [_contentView updateScrollViewHeight];
-}
-
 - (CGFloat)contentHeight {
   return _contentView ? _contentView.frame.size.height : 0;
 }
@@ -141,22 +133,8 @@ using namespace facebook::react;
   return CGRectGetMaxY([_peekView convertRect:_peekView.bounds toView:_contentView]);
 }
 
-- (void)layoutFooter {
-  if (_footerView) {
-    CGFloat height = _footerView.frame.size.height;
-    if (height > 0) {
-      [_footerView setupConstraintsWithHeight:height];
-    }
-  }
-}
-
 - (void)updateFooterKeyboardOffset {
   [_footerView applyKeyboardOffset];
-}
-
-- (void)setScrollableEnabled:(BOOL)scrollableEnabled {
-  _scrollableEnabled = scrollableEnabled;
-  _scrollableSet = YES;
 }
 
 - (void)setScrollableOptions:(ScrollableOptions *)scrollableOptions {
@@ -165,12 +143,12 @@ using namespace facebook::react;
 }
 
 - (void)setupScrollable {
-  if (_scrollableSet && _contentView) {
+  if (_contentView) {
     CGFloat bottomInset = 0;
     if (_insetAdjustment == TrueSheetViewInsetAdjustment::Automatic) {
       bottomInset = [WindowUtil keyWindow].safeAreaInsets.bottom;
     }
-    [_contentView setupScrollable:_scrollableEnabled bottomInset:bottomInset];
+    [_contentView setupScrollableWithBottomInset:bottomInset];
     [_contentView applyScrollEdgeEffects:_scrollableOptions];
     if (@available(iOS 26.0, *)) {
       [self setupEdgeInteractions];
@@ -267,12 +245,6 @@ using namespace facebook::react;
 
 - (void)updateProps:(Props::Shared const &)props oldProps:(Props::Shared const &)oldProps {
   [super updateProps:props oldProps:oldProps];
-}
-
-#pragma clang diagnostic ignored "-Wobjc-missing-super-calls"
-- (void)updateLayoutMetrics:(const LayoutMetrics &)layoutMetrics
-           oldLayoutMetrics:(const LayoutMetrics &)oldLayoutMetrics {
-  // Intentionally skip super - AutoLayout handles container's frame, not Yoga
 }
 
 #pragma mark - TrueSheetContentViewDelegate