+++ src/TrueSheet.types.ts
@@ -461,17 +461,6 @@ export interface TrueSheetProps extends ViewProps {
   footerStyle?: StyleProp<ViewStyle>;
 
   /**
-   * On iOS, automatically pins ScrollView or FlatList to fit within the sheet's available space.
-   * When enabled, the ScrollView's top edge will be pinned below any top sibling views,
-   * and its left, right, and bottom edges will be pinned to the container.
-   *
-   * On Android, it adds additional style to the content for scrollable to work.
-   *
-   * @default false
-   */
-  scrollable?: boolean;
-
-  /**
    * Options for scrollable behavior.
    */
   scrollableOptions?: ScrollableOptions;